build-sys: Add check-lib target
The referenced commits created the library infrastructure and test
program to validate that the structures and macros line up with
each other.

The library needs to be (re)built with -DITEMTABLE_DEBUG and then
the test program ran.  We clean before and after so we are not
testing a non-debug library or having a debug library hanging around
to cause future problems.

Due to test_Itemtables depending on the library, we don't need to
explicitly build the library.

To validate the library structure/header corrospondence run:
make check-lib

References:
 commit e616409aa497b5dc656dc7a5bd0a21adf0ee4d36
 commit 92d0297e1e4d5946c5b098e37c91c7e524a0eca0
 https://www.freelists.org/post/procps/keep-on-patchin,19