top: revert part of a patch regarding separate threads
In the patch referenced below, some logic was moved in
that frame_make function in an effort to overlap those
background threads with additional main thread duties.

Unfortunately, the auto-size column feature was broken
in the process. So this patch will revert a portion of
the earlier patch so as to restore the broken feature.

Reference(s):
. Sep 2021, top overlap additional processing
commit f11f43b4f7118f284f9aec19f6885d163848ffe7

Signed-off-by: Jim Warner <james.warner@comcast.net>