top: add visual clue when focus toggle ('F') is active
When in forest view mode, that focus toggle ('F') is a
useful tool occasionally. But, if a focused parent has
enough cloned siblings to exceed screen rows, it could
be hard to remember that such a toggle remains active.

So, this patch will provide a subtle visual clue added
to the leftmost position in the COMMAND column. Now if
the focus toggle was active, regardless of total tasks
affected, the users will always know when it's active.

Reference(s):
. 07/28/21, introduced new focus toggle ('F')
commit bdd50fa2fc00edb51f6125334d2329fdce12c4ba
. 09/23/21, ensure focused tasks stay focused
commit d7e6c27a7944fbaa1493f207fd521bd59984c955

Signed-off-by: Jim Warner <james.warner@comcast.net>