top: add visual clue when focus toggle ('F') is active
When in forest view mode, that focus toggle ('F') is a
useful tool occasionally. But, if a focused parent has
enough cloned siblings to exceed screen rows, it could
be hard to remember that such a toggle remains active.

So, this patch will provide a subtle visual clue added
to the leftmost position in the COMMAND column. Now if
the focus toggle was active, regardless of total tasks
affected, the users will always know when it's active.

Reference(s):
. -7/24/21, introduced new focus toggle
commit 3e922e671daf09ae60ee211b5c0ef94996ee573d
. 09/23/21, ensure focused tasks stay focused
commit 69978e365043f27305e487709474947bb377084d

Signed-off-by: Jim Warner <james.warner@comcast.net>